fix latent warnings
We're removing various warning suppressions by default from the wrapper
(crrev.com/c/6672951); this CL fixes up new -Werrors that result on
zephyr builders.

BUG=b:296092419
TEST=CQ+1 w/ disable CL, ran the build_api endpoint locally

+++ zephyr/emul/emul_rt1739.c
@@ -91,7 +91,7 @@ static void add_set_reg_history_entry(struct rt1739_data *data, int reg,
 	sys_slist_append(&data->set_private_reg_history, &entry->node);
 }
 
-static int rt1739_emul_read(const struct emul *emul, int reg, uint16_t *val,
+static int rt1739_emul_read(const struct emul *emul, int reg, uint8_t *val,
 			    int bytes, void *unused_data)
 {
 	struct rt1739_data *data = emul->data;
